SSCS-10268 added tests + refactoring
