+++ build.gradle
@@ -315,6 +315,8 @@ dependencies {
 
   implementation group: 'org.slf4j', name: 'slf4j-api', version: versions.slf4j
 
+  implementation group: 'com.networknt', name: 'json-schema-validator', version: '1.0.66'
+
   testImplementation group: 'io.rest-assured', name: 'rest-assured', version: versions.restAssured
   testImplementation group: 'io.rest-assured', name: 'xml-path', version: versions.restAssured
   testImplementation group: 'io.rest-assured', name: 'json-path', version: versions.restAssured
