+++ src/main/java/uk/gov/hmcts/reform/sscs/model/servicehearingvalues/CaseFlags.java
@@ -1,11 +1,14 @@
-package uk.gov.hmcts.reform.sscs.model;
+package uk.gov.hmcts.reform.sscs.model.servicehearingvalues;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
+import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+
+import java.util.List;
 
 @Data
 @Builder(toBuilder = true)
@@ -13,8 +16,9 @@ import lombok.NoArgsConstructor;
 @JsonIgnoreProperties(ignoreUnknown = true)
 @AllArgsConstructor
 @NoArgsConstructor
-public class HearingWindowDateRange {
+public class CaseFlags {
 
-    private String hearingWindowStartDateRange;
-    private String hearingWindowEndDateRange;
+    private List<PartyFlags> flags;
+    @JsonProperty("flagAmendURL")
+    private String flagAmendUrl;
 }
