+++ src/main/java/uk/gov/hmcts/reform/sscs/model/servicehearingvalues/DOWUnavailabilityType.java
@@ -1,8 +1,9 @@
-package uk.gov.hmcts.reform.sscs.model;
+package uk.gov.hmcts.reform.sscs.model.servicehearingvalues;
 
 import lombok.Getter;
 
 @Getter
+@SuppressWarnings("checkstyle:abbreviationaswordinname")
 public enum DOWUnavailabilityType {
 
     AM("AM"),
