+++ src/main/java/uk/gov/hmcts/reform/sscs/model/servicehearingvalues/HearingLocation.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.sscs.model;
+package uk.gov.hmcts.reform.sscs.model.servicehearingvalues;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
@@ -7,16 +7,16 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.util.List;
-
 @Data
 @Builder(toBuilder = true)
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @JsonIgnoreProperties(ignoreUnknown = true)
 @AllArgsConstructor
 @NoArgsConstructor
-public class ScreenNavigation {
-    private String screenName;
-    private String conditionKey;
-    private List<Navigation> navigation;
+public class HearingLocation {
+
+    private String locationId;
+    private String locationName;
+    private String locationType;
+    private String region;
 }
