+++ src/main/java/uk/gov/hmcts/reform/sscs/model/servicehearingvalues/Navigation.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.sscs.model;
+package uk.gov.hmcts.reform.sscs.model.servicehearingvalues;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
@@ -13,8 +13,9 @@ import lombok.NoArgsConstructor;
 @JsonIgnoreProperties(ignoreUnknown = true)
 @AllArgsConstructor
 @NoArgsConstructor
-public class RelatedParties {
+public class Navigation {
 
-    private String relatedPartyID;
-    private String relationshipType;
+    private String conditionOperator;
+    private String conditionValue;
+    private String resultValue;
 }
