+++ src/main/java/uk/gov/hmcts/reform/sscs/model/servicehearingvalues/OrganisationDetails.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.sscs.model;
+package uk.gov.hmcts.reform.sscs.model.servicehearingvalues;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
@@ -7,20 +7,15 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.util.List;
-
 @Data
 @Builder(toBuilder = true)
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @JsonIgnoreProperties(ignoreUnknown = true)
 @AllArgsConstructor
 @NoArgsConstructor
-public class Judiciary {
+public class OrganisationDetails {
 
-    private List<String> roleType;
-    private List<String> authorisationTypes;
-    private List<String> authorisationSubType;
-    private List<PanelComposition> panelComposition;
-    private List<PanelPreference> judiciaryPreferences;
-    private List<String> judiciarySpecialisms;
+    private String name;
+    private String organisationType;
+    private String cftOrganisationID;
 }
