+++ src/main/java/uk/gov/hmcts/reform/sscs/model/servicehearingvalues/PartyFlags.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.sscs.model;
+package uk.gov.hmcts.reform.sscs.model.servicehearingvalues;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
@@ -13,8 +13,11 @@ import lombok.NoArgsConstructor;
 @JsonIgnoreProperties(ignoreUnknown = true)
 @AllArgsConstructor
 @NoArgsConstructor
-public class UnavailabilityRange {
+public class PartyFlags {
 
-    private String unavailableFromDate;
-    private String unavailableToDate;
+    private String partyName;
+    private String flagParentId;
+    private String flagId;
+    private String flagDescription;
+    private String flagStatus;
 }
