+++ src/main/java/uk/gov/hmcts/reform/sscs/model/servicehearingvalues/UnavailabilityDOW.java
@@ -1,7 +1,8 @@
-package uk.gov.hmcts.reform.sscs.model;
+package uk.gov.hmcts.reform.sscs.model.servicehearingvalues;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
+import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -13,8 +14,11 @@ import lombok.NoArgsConstructor;
 @JsonIgnoreProperties(ignoreUnknown = true)
 @AllArgsConstructor
 @NoArgsConstructor
-public class PanelComposition {
+@SuppressWarnings("checkstyle:abbreviationaswordinname")
+public class UnavailabilityDOW {
 
-    private String memberType;
-    private Integer count;
+    @JsonProperty("DOW")
+    private DOW dow;
+    @JsonProperty("DOWUnavailabilityType")
+    private DOWUnavailabilityType dowUnavailabilityType;
 }
